add pipeline jobs and improve readme
