+++ .github/workflows/cicd.yml
@@ -9,39 +9,86 @@ on:
       - main  # Run on pull requests to the main branch
 
 jobs:
-  lint-test:
+  # Job 1: Linting with flake8
+  lint-flake8:
     runs-on: ubuntu-latest
-
     steps:
-      # Step 1: Check out the code
       - name: Checkout code
         uses: actions/checkout@v3
 
-      # Step 2: Set up Python environment
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: "3.9"  # Change to your project's Python version
+          python-version: "3.9"
 
-      # Step 3: Install dependencies
-      - name: Install dependencies
+      - name: Install flake8
         run: |
           python -m pip install --upgrade pip
-          pip install flake8 isort mypy pytest  # Add other dependencies as needed
-          pip install -r requirements-dev.txt  # If you use a requirements file
+          pip install flake8
 
-      # Step 4: Run flake8
       - name: Run flake8
-        run: flake8 trace_logger/ tests/  # Adjust the target directory if needed
+        run: flake8 trace_logger/ tests/
 
-      # Step 5: Run isort check
+  # Job 2: Import sorting with isort
+  lint-isort:
+    runs-on: ubuntu-latest
+    needs: lint-flake8
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.9"
+
+      - name: Install isort
+        run: |
+          python -m pip install --upgrade pip
+          pip install isort
+
       - name: Run isort check
         run: isort trace_logger/ tests/ --check-only
 
-      # Step 6: Run mypy
-      - name: Run mypy
-        run: mypy trace_logger/ tests/  # Adjust the target directory if needed
+  # Job 3: Static type checking with mypy
+  lint-mypy:
+    runs-on: ubuntu-latest
+    needs: lint-isort
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
 
-      # Step 7: Run tests
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.9"
+
+      - name: Install mypy
+        run: |
+          python -m pip install --upgrade pip
+          pip install mypy
+
+      - name: Run mypy
+        run: mypy trace_logger/ tests/
+
+  # Job 4: Run tests with pytest
+  run-tests:
+    runs-on: ubuntu-latest
+    needs: lint-mypy
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.9"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install pytest
+          pip install -r requirements-dev.txt  # If you have a dev requirements file
+
       - name: Run tests
-        run: pytest tests/ -vv  # Adjust the command if needed (e.g., specify test directory)
+        run: pytest tests/ -vv
