+++ secure_logs/config.py
@@ -6,30 +6,6 @@ import re
 
 from .utils import get_trace_id
 
-_LOGGING_CONFIGURED = False
-
-
-def configure_logging(level: str, sensitive_patterns: list[str] = None):
-    """
-    Configures the logging level for the library and applies sensitive value redaction.
-
-    Args:
-        level (str): The logging level (e.g., DEBUG, INFO, WARNING, ERROR, CRITICAL).
-        sensitive_patterns (list[str]): List of regex patterns for sensitive values.
-    """
-    global _LOGGING_CONFIGURED
-    if not _LOGGING_CONFIGURED:
-        logging.basicConfig(
-            level=getattr(logging, level.upper(), logging.INFO),
-            format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-        )
-        if sensitive_patterns:
-            sensitive_filter = SensitiveValueFilter(sensitive_patterns)
-            logging.getLogger().addFilter(sensitive_filter)
-
-        _LOGGING_CONFIGURED = True
-
-
 class TraceIDAdapter(logging.LoggerAdapter):
     def process(self, msg, kwargs):
         trace_id = get_trace_id()
@@ -40,36 +16,6 @@ class TraceIDAdapter(logging.LoggerAdapter):
             f"[trace_id: {trace_id}] [function: {function_name}] {msg}",
             kwargs,
         )  # noqa
-
-
-def get_logger(
-    name: str = __name__, sensitive_patterns: list[str] = None, show_last: int = 0
-) -> TraceIDAdapter:
-    """
-    Returns a TraceIDAdapter logger with the given name. Optionally hides sensitive values.
-
-    Args:
-        name (str): The name of the logger.
-        sensitive_patterns (list[str]): List of regex patterns for sensitive values.
-
-    Returns:
-        TraceIDAdapter: The logger instance.
-    """
-    logger = logging.getLogger(name)
-    if not logging.getLogger().hasHandlers():
-        configure_logging(
-            level="DEBUG", sensitive_patterns=sensitive_patterns, show_last=show_last
-        )
-        logger.warning(
-            "No logging configuration detected. Consider calling "
-            "configure_logging()."
-        )
-
-    if sensitive_patterns:
-        sensitive_filter = SensitiveValueFilter(sensitive_patterns, show_last)
-        logger.addFilter(sensitive_filter)
-
-    return TraceIDAdapter(logger, {})
 
 
 class SensitiveValueFilter(logging.Filter):
