+++ secure_logs/middleware.py
